refactoring with sub-folders and sub commands
