+++ build.sh
@@ -3,4 +3,4 @@
 podman run --rm -v $PWD:/src \
     --workdir /src/ \
     registry.suse.com/bci/golang:latest \
-    go build -o . ./...
+    go build -o ./bin ./...
