--- cmd/uyunictl/main.go
@@ -1,9 +1,0 @@
-package main
-
-import (
-	ctl "github.com/uyuni-project/uyuni-tools/internal/uyunictl"
-)
-
-func main() {
-	ctl.Execute()
-}
