+++ shared/types/global.go
@@ -1,0 +1,5 @@
+package types
+
+type GlobalFlags struct {
+	Verbose bool
+}
